Group top20 components in folder & clean up code
