+++ client/components/top20/top20Wrapper.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Console from './Top20Console';
 import Section from 'grommet/components/Section';
 import Header from 'grommet/components/Header';
 import Title from 'grommet/components/Title';
-import HeaderContainer from '../containers/headerContainer';
+import Console from './top20Console';
+import HeaderContainer from '../../containers/headerContainer';
 
 
 const Top20Wrapper = () =>
@@ -12,17 +12,12 @@ const Top20Wrapper = () =>
       <Header className="landing-main-nav" size="small" float={false} fixed={true}>
         <HeaderContainer />
       </Header>
-      <Header
-        className="top20-header header"
-        size="small"
-      >
+      <Header className="top20-header header" size="small" >
         <Title>
           Campaign Finance Top 20 Lists
         </Title>
       </Header>
-      <Section
-        className="top20-main"
-      >
+      <Section className="top20-main" >
         <Console />
       </Section>
     </div>
